requirements.txt
python-telegram-bot==13.11
numpy==1.22.3
scikit-learn==1.0.2
pandas==1.4.2


